kayunkim/lectures/Vue.js/practice/0608/Zoom/workshop/articleBack/requirements.txt
Django==2.1.15
djangorestframework==3.11.0
pytz==2020.1
